pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "context-engine"
version = "0.1.0"
description = "Cognitive session reconstruction engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyobjc-framework-quartz>=12.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["context_engine*"]
exclude = ["data*", "docs*", "notebooks*", "scripts*", "tests*"]

[tool.pytest.ini_options]
pythonpath = ["."]
